✨ setup cli - create file structure
now the cli can ensure that the necessary file structure is in
place to allow the rest of the functions to work. the one part
that remains to do is _saving_ the targetDir to a local environment
variable that the rest of the application will have access to.

that will ensure that the state is not lost over time (at least not
until setup is run again).