✨ notes - published and counts
this commit adds two pieces of new functionality:
1. finding the date of the most recent, latest, and first published
note
2. finding the count of notes by date written (date) and published
(publish). the latter is the default.

It also includes significant refactoring - specifically around the
frontmatter, which is now all lowercase. this keeps it inline with
the frontmatter interface and makes the helper functions behave as
expected.

date helpers for formatting also encourage a human readable, albeit
fixed, way of seeing the dates. this can be improved later by using
the `DEFAULT_DATE_FORMAT` in the config.