✨ update a note
the first part of updating a note is now complete - updating the
actual note's frontmatter. this can be done declaratively through
the use of flags or interactively.

currently declarative is the default, which means if no flags are
set, the note will "update" but won't change.

next step is to update the `.contents` file. notes for what is
needed have been added.