🧰 set up np for publishing [nom-37]
🧰 tooling updates
- move from `build` to `dist`
- update details in the package.json, including repository,
issue tracker, keywords
- add a test script placeholder to avoid bombing on publish
- add `.npmignore` (see note below)
- add `@babel/runtime` and `@babel/plugin-transform-runtime` to dependencies

📝 docs updates
- revamp README and move gifs to cloudinary

note re: `.npmignore`

when publishing, we need the `dist` directory to be included.
by default, npm looks at the `.gitignore` file. however, for our
purposes, that was too restrictive.

by adding a `.npmignore` file, we can specify only what we want
npm to ignore.

in our case, the defaults are sufficient, and so a blank file is
all that's necessary not to fall back to the `.gitignore`.

for more information see: https://docs.npmjs.com/misc/developers#keeping-files-out-of-your-package

note re: @babel/runtime and @babel/plugin-transform-runtime

unlike most babel plugins, the transform-runtime _is_ required
in the transpiled code.

this move will make sure that it is part of the dependencies that
are installed on download.

references:
- https://stackoverflow.com/a/55133339
- https://github.com/babel/babel/discussions/11611