🐛 fix nom update date bug [nom-47]
simplify how dates are set within `nom` to basically just use the
default date format. this means that dates don't move around the
same way and validation is much more reliable. it _does_ mean that
typically dates should not be written manually, otherwise, it's
quite likely that the timezone will be off and the final date will be
off.

it's also possible that this will introduce a bug into the dates
command.