🏗️🐛  remove / update actions + remove custom fm
custom frontmatter has been removed for now.

parseOptions was refactored to use the frontmatter object, not a
map of frontmatter keys.

remove note action was updated to ensure that both the file and the
content row existed prior to deleting either

a new helper function for generating the title (i.e. the md link for
.contents) was added

a lot of refactoring of note helpers to be part of the notes class.