🐛 fix silent errors
error messages were being eaten and never presented to the user to
help them debug the issue.

this adds a new standardization around how errors should be printed
and generated.

the "big picture" is that at the top of the call stack we need to
actually _print_ the errors. we can't always catch and throw a new
one.

The simple rule of thumb is that actions should print errors,
utility functions should throw them. Organization isn't always
perfect though.