✨ comment out undefined frontmatter [NOM-62]
if a note is created, but without frontmatter, that should result
in the line for the frontmatter being created, but commented out.

this is already what was happening with certain fields, such as
tags and categories.

this commit now makes it universal by checking _first_ to see if
the value of the frontmatter exists, _then_ processing based on the
key itself. since keys can have special processing, this avoids a
class of bugs where that special processing had a silent default,
as was the case with dates.